♻️ Every iteration block draws its next from a queue
We had a problem where we were getting a "Generator is already
running" error because we had two loops that were driving the
iterator; The first loop drove it until it was either done or
interrupted, and the second loop drove it until all instructions were
exhausted after such time. However, because all the continuations are
inherently synchronous, there were cases where a continuation invoked
from within `getNext()` was causing the second loop to be entered
which then called `next()` on the same iterator.

This changes every block of instructions to use a single `while` loop,
and instead draw its calls to next from a queue.

In order for this to work, the `Observable#observe()` had to be made
stateful. That way, any events that are delivered while not explicitly
listening, are not dropped.

This allowed us not only to condense to a single loop, but drastically
reduce the complexity of the `Block` interface and the events that it
emits. Now, instead of emitting a `instruction`, `exit`, and
`exhaustion` events, it emits a single `BlockResult`

This, combined with the new stateful queues, means that the `action`
instruction implementation was able to be simplified greatly because
`resolve` and `reject` just push values onto a queue, rather than
changing the in-memory references to a "return" function.