Use tail-optimized continuations
In Effection, we don't actually use the return value of any of the
delimited continuations. Because of that, we can avail ourselves of
the tail-call optimizations that were introduced here
https://github.com/thefrontside/continuation/pull/11

This will cause all synchronous continuations to evaluate on the same
run loop. Before, this would cause a stack over flow.

```ts
for (let i = 0; i < 1_000_00; i++;) {
  yield* action(function*(resolve) { resolve() });
}
```

Now, each action continuation happens on the same continuation
reduction loop, and so results in another iteration rather than
recursion.