♻️Remove event stream entirely from action flow
Errors were not being propagates to an action when its frame crashed
because of a resource and not because there was something wrong with
its body or its `reject()` continuation was invoked.

It appears that this is a bug in `EventStream` where if close ends up
calling itself, then you get a "double close" race error. However, now
that `continuation` is quasi tail recursive, I decided to see if we
could get by just using that, and it works because all continuations
are inherently queued.