✨Add `call()` operation
In `v2`, you could `yield` directly to a generator function, and that
was the way to run an completely encapsulated anonymous
task.

```ts
function run* function() {
  yield function*() {
    //do something
  }
  yield function*() {
    //do something else.
  }
}
```

However, in `v3`, rather than a union of [seven distinct types][1],
there is only a single `Operation` type, and `() => Generator` is not
one of them. This simplifies internals greatly, However, we still need a way
to invoke an anonymous generator functions without having to define it
elsewhere and call it, or use an IIFE to get a hold of the iterable.

```ts
function run* function() {
  yield* (function*() {
    //do something
  })();
  yield* (function*() {
    //do something else.
  })();
}
```

With the `call()` operation, you can pass the anonymous generator
function, and it will be wrapped in an action to run it.

```ts
function run* function() {
  yield* call(function*() {
    //do something
  });
  yield* call(function*() {
    //do something else.
  });
}
```

Also, there is another concern that is only germaine to `v3` which is
that `yield*` only "inlines" instructions into the current block. It
does not create a new `Frame`. That means that any resources will be
added to the frame running the instructions, and it also means that
any thrown errors will crash that same frame.

`call()` is implemented with an action which creates its own frame,
that way, any resources allocated within the scope of the generator
function will be cleaned up before the call returns, and any errors
raised inside it can be caught.

[1]: https://github.com/thefrontside/effection/blob/v2/packages/core/src/operation.ts#L55-L62